+++ extension/worker_wrapper.js
@@ -1,0 +1,7 @@
+try {
+    importScripts("src/mv3_polyfill.js", "src/configuration.js", "src/utils.js", "src/background.js");
+  } catch (e) {
+    console.error(e);
+  }
+    
+